Split fourier.hpp between header and source
